Add a sidebar to the hub windows (#1045250)
This adds a sidebar to the left of the hub windows. The reason for
this is to visually reinforce the left-to-right linear nature of the
hub windows. There is a variable (sidebar_width_percentage) you can
set to control the percentage width of the screen that the sidebar
occupies - I've set it to 18% which seems reasonable. It auto-centers
the logo on top of the sidebar as well. There are three layers to
the sidebar: the base color, a texture pixbuf, and the logo pixbuf
on top.