Highlight done buttons via 'suggested-action' style class (#1045250)
Call more attention to the 'Done' buttons in the upper left of spoke
windows by adding the 'suggested-action' button style class to those
buttons. The effect will be buttons with a blue background. This
should help draw more attention to the buttons for users who have
a hard time finding them. Reference to GNOME docs on usage of this
button style are here: https://wiki.gnome.org/HowDoI/Buttons