Add a sidebar to the standalone and hub windows (#1045250)
This adds a sidebar to the left of the standalone and hub windows. The
reason for this is to visually reinforce the left-to-right linear nature
of the standalone windows. There is a variable (sidebar_width_percentage)
you can set to control the percentage width of the screen that the
sidebar occupies - I've set it to 18% which seems reasonable. It auto-
centers the logo on top of the sidebar as well. The sidebar is controlled
by the 'sidebar' class in the css and the 'logo' class in the css. In
this patch, the sidebar is set to have a grey background with a
partially-transparent image texture layered on top, and a logo layered
on top of that, centered horizontally 20px down from the top of the
sidebar. To restyle the sidebar, you only have to change the
relevant section in the anaconda-gtk.css file.